telegram_bot/requirements.txt
aiofiles==24.1.0
aiogram==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.16.4
annotated-types==0.7.0
APScheduler==3.11.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
Deprecated==1.2.18
distro==1.9.0
frozenlist==1.7.0
greenlet==3.2.4
idna==3.10
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.6.3
netaddr==1.3.0
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
SQLAlchemy==2.0.42
typing-inspection==0.4.1
typing_extensions==4.14.1
tzlocal==5.3.1
urllib3==2.5.0
wrapt==1.17.3
yarl==1.20.1
yookassa==3.6.0
